- Editado função de atualizacao automatica - Inserido Arquivo para importar os eventos dos pdfs

+++ update.py
@@ -4,8 +4,7 @@ import sys
 import subprocess
 from urllib.request import urlretrieve
 from config_json import ConfigManager
-import tkinter as tk
-from tkinter import messagebox
+
 
 GITHUB_RELEASES_API_URL = 'https://api.github.com/repos/Patoxzor/importpdf/releases/latest'
 
@@ -14,45 +13,46 @@ def get_latest_release_info():
     return response.json() if response.ok else None
 
 def download_and_install_update(url, version):
+    # Baixando o executável
     local_filename = f'Mixpdf - {version}.exe'
-
     urlretrieve(url, local_filename)
 
+    # Preparando para instalar a atualização
     current_app_path = os.path.abspath(sys.argv[0])
-    old_app_path = f'{current_app_path}_old'
 
-    if os.path.exists(old_app_path):
-        os.remove(old_app_path)
-    os.rename(current_app_path, old_app_path)
-    os.replace(local_filename, current_app_path)
-    ConfigManager.save_version(version)
-    subprocess.Popen([current_app_path])
-    sys.exit()
+    # Verificar se o caminho atual é um arquivo .exe
+    if current_app_path.endswith('.exe'):
+        old_app_path = f'{current_app_path}_old'
+        if os.path.exists(old_app_path):
+            os.remove(old_app_path)
+        os.rename(current_app_path, old_app_path)
+        os.replace(local_filename, current_app_path)
 
-def notify_user_of_update(latest_version):
-    root = tk.Tk()
-    root.withdraw()
-    response = messagebox.askyesno("Atualização Disponível", 
-                                   f"Uma nova versão {latest_version} está disponível. Deseja atualizar agora?")
-    if response:
-        return True
+        # Atualizando a versão registrada
+        ConfigManager.save_version(version)
+
+        # Reiniciando a aplicação
+        subprocess.Popen([local_filename])
+        sys.exit()
     else:
-        return False
+        # Atualização para ambiente de desenvolvimento, não reiniciar
+        print("Atualização baixada. Reinicie manualmente o aplicativo.")
+
 
 def check_for_updates():
     current_version = ConfigManager.load_version()
     latest_release = get_latest_release_info()
-    
+
     if latest_release:
         latest_version = latest_release['tag_name'].replace('V', '')
+        assets = latest_release['assets']
+        browser_download_url = None
 
-        # Comparando a versão atual com a versão mais recente
-        if latest_version != current_version:
-            print(f'Nova versão disponível: {latest_version}')
-            assets = latest_release['assets']
-            if assets:
-                browser_download_url = assets[0]['browser_download_url']
-                download_and_install_update(browser_download_url, latest_version)
-        else:
-            print("Você já está com a versão mais recente.")
+        for asset in assets:
+            if 'Mixpdf' in asset['name']:
+                browser_download_url = asset['browser_download_url']
 
+        if latest_version != current_version and browser_download_url:
+            return latest_version, browser_download_url
+
+    return None, None
